tegra_fb: re-enable double-buffering support
Commit 185b396 made
double-buffering support in the fb driver conditional
on CONFIG_ANDROID, but was an incomplete removal of the
feature - userland apps can still set up the device
for double-buffering, since it looks like it still supports
it, but then get EINVAL errors when tryng to use it.

Since the difference in memory is only 32MiB, just remove
the CONFIG_ANDROID check and enable double-buffering
in all builds.

Signed-off-by: Matt Madison <matt@madison.systems>