tegra: host: t210: revert to per channel policy
Commit 2342da5679501364272d6c1263d6405afc45006d claims that MLOCK host1x feature is not supported in T210.

This contradicts TRM.

In addition to that, this change produces massive tearing on applications that use video hw acceleration via NVDEC.

Revert the resource_policy changes to fix the issue but retain the rest of the changes.
That's in case the MLOCK support was actually artificially dropped in latest NVDEC/NVENC/NVJPG/VIC firmware.